#5828 - Refactoring
