+++ sormas-ui/src/main/java/de/symeda/sormas/ui/caze/CaseDataView.java
@@ -46,7 +46,7 @@ import de.symeda.sormas.ui.utils.CssStyles;
 import de.symeda.sormas.ui.utils.DetailSubComponentWrapper;
 import de.symeda.sormas.ui.utils.LayoutUtil;
 import de.symeda.sormas.ui.utils.ViewMode;
-import de.symeda.sormas.ui.utils.components.SideComponentLayout;
+import de.symeda.sormas.ui.utils.components.sidecomponent.SideComponentLayout;
 
 /**
  * CaseDataView for reading and editing the case data fields. Contains the
